kafka_worker/requirements.txt
faust
sqlalchemy
python-dotenv